More refactoring. Added constants file.
