+++ mchartanalyzer/constants.py
@@ -1,0 +1,5 @@
+import os
+
+MODULE_DIR = os.path.dirname(os.path.realpath(__file__))
+PROJECT_DIR = os.path.dirname(MODULE_DIR)
+DATA_OUTPUT_DIR = os.path.join(PROJECT_DIR, "mchartanalyzer-data")
