+++ mchartanalyzer/objects/artistdata.py
