+++ mchartanalyzer/objects/chartdata.py
